+++ packages/file_picker_linux/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.1
+
+- Relax `dbus` to `^0.7.13` so dependents can resolve `xml` 7. [#2130](https://github.com/miguelpruivo/flutter_file_picker/issues/2130)
+
 ## 1.0.0
 
 - Initial release of Linux implementation package for `file_picker`.